Function_App/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas==1.5.1
numpy==1.22.4
streamlit==1.11.0
scikit-learn==1.1.3
requests
